OrderWithdrawalFormHandler logic
