+++ plugins/woocommerce/templates/myaccount/form-order-withdrawal.php
@@ -30,6 +30,7 @@ $action_field            = isset( $action_field ) ? (string) $action_field : '';
 $action_review           = isset( $action_review ) ? (string) $action_review : '';
 $action_confirm          = isset( $action_confirm ) ? (string) $action_confirm : '';
 $action_edit             = isset( $action_edit ) ? (string) $action_edit : '';
+$form_action_url         = isset( $form_action_url ) ? (string) $form_action_url : '';
 $shop_url                = isset( $shop_url ) ? (string) $shop_url : wc_get_page_permalink( 'shop' );
 $form_field_keys         = array( 'first_name', 'last_name', 'email', 'email_confirmation', 'order_number' );
 $withdrawal_type_field   = $fields['withdrawal_type'] ?? array();
@@ -51,25 +52,25 @@ $secondary_button_class = implode( ' ', array_merge( $button_classes, array( 'wo
 	<h2><?php esc_html_e( 'Withdraw from contract', 'woocommerce' ); ?></h2>
 
 	<?php if ( 'confirmation' === $screen ) : ?>
-		<p><strong><?php esc_html_e( 'Your withdrawal has been submitted.', 'woocommerce' ); ?></strong></p>
+		<p><strong><?php esc_html_e( 'Your withdrawal details are ready.', 'woocommerce' ); ?></strong></p>
 		<p class="woocommerce-order-withdrawal-content__note">
 			<?php
 			printf(
 				/* translators: %s: Email address. */
-				esc_html__( 'We\'ve emailed an acknowledgment to %s with your details and the date and time of submission. Keep it as proof of your withdrawal.', 'woocommerce' ),
+				esc_html__( 'Nothing has been sent yet. In a future version, we\'ll email an acknowledgment to %s with your details and the date and time of submission.', 'woocommerce' ),
 				esc_html( $data['email'] ?? '' )
 			);
 			?>
 		</p>
-		<p class="woocommerce-order-withdrawal-content__note"><?php esc_html_e( 'We\'ll review your request and contact you about next steps, including any refund due.', 'woocommerce' ); ?></p>
+		<p class="woocommerce-order-withdrawal-content__note"><?php esc_html_e( 'We are still preparing the submission flow for this form.', 'woocommerce' ); ?></p>
 		<p><a href="<?php echo esc_url( $shop_url ); ?>"><?php esc_html_e( 'Back to the shop', 'woocommerce' ); ?></a></p>
 	<?php elseif ( 'review' === $screen ) : ?>
 		<p class="woocommerce-order-withdrawal-content__intro"><?php esc_html_e( 'Check your details before confirming.', 'woocommerce' ); ?></p>
-		<p class="woocommerce-order-withdrawal-content__note"><?php esc_html_e( 'Nothing has been sent yet. Your withdrawal is submitted when you select "Confirm withdrawal".', 'woocommerce' ); ?></p>
+		<p class="woocommerce-order-withdrawal-content__note"><?php esc_html_e( 'Nothing has been sent yet. Confirming only completes this prototype flow.', 'woocommerce' ); ?></p>
 
 		<ol class="woocommerce-order-withdrawal-content__steps" aria-label="<?php esc_attr_e( 'Order withdrawal progress', 'woocommerce' ); ?>">
-			<li class="woocommerce-order-withdrawal-content__step"><?php esc_html_e( '1. Your details', 'woocommerce' ); ?></li>
-			<li class="woocommerce-order-withdrawal-content__step woocommerce-order-withdrawal-content__step--current" aria-current="step"><?php esc_html_e( '2. Review and confirm', 'woocommerce' ); ?></li>
+			<li class="woocommerce-order-withdrawal-content__step"><?php esc_html_e( 'Your details', 'woocommerce' ); ?></li>
+			<li class="woocommerce-order-withdrawal-content__step woocommerce-order-withdrawal-content__step--current" aria-current="step"><?php esc_html_e( 'Review and confirm', 'woocommerce' ); ?></li>
 		</ol>
 
 		<dl class="woocommerce-order-withdrawal-content__review">
@@ -85,13 +86,13 @@ $secondary_button_class = implode( ' ', array_merge( $button_classes, array( 'wo
 			<?php
 			printf(
 				/* translators: %s: Email address. */
-				esc_html__( 'After you confirm, we\'ll email an acknowledgment to %s with these details and the date and time of submission. Keep it as proof of your withdrawal.', 'woocommerce' ),
+				esc_html__( 'Nothing will be sent in this version. In a future version, we\'ll email an acknowledgment to %s with these details and the date and time of submission.', 'woocommerce' ),
 				esc_html( $data['email'] ?? '' )
 			);
 			?>
 		</p>
 
-		<form class="woocommerce-OrderWithdrawalForm" method="post" novalidate>
+		<form class="woocommerce-OrderWithdrawalForm" action="<?php echo esc_url( $form_action_url ); ?>" method="post" novalidate>
 			<?php foreach ( $hidden_fields as $hidden_field ) : ?>
 				<input type="hidden" name="<?php echo esc_attr( $hidden_field['name'] ?? '' ); ?>" value="<?php echo esc_attr( $hidden_field['value'] ?? '' ); ?>" />
 			<?php endforeach; ?>
@@ -107,31 +108,25 @@ $secondary_button_class = implode( ' ', array_merge( $button_classes, array( 'wo
 		<p class="woocommerce-order-withdrawal-content__note"><?php esc_html_e( 'Some items, like personalized products, may not be eligible. We review every request and reply by email.', 'woocommerce' ); ?></p>
 
 		<ol class="woocommerce-order-withdrawal-content__steps" aria-label="<?php esc_attr_e( 'Order withdrawal progress', 'woocommerce' ); ?>">
-			<li class="woocommerce-order-withdrawal-content__step woocommerce-order-withdrawal-content__step--current" aria-current="step"><?php esc_html_e( '1. Your details', 'woocommerce' ); ?></li>
-			<li class="woocommerce-order-withdrawal-content__step"><?php esc_html_e( '2. Review and confirm', 'woocommerce' ); ?></li>
+			<li class="woocommerce-order-withdrawal-content__step woocommerce-order-withdrawal-content__step--current" aria-current="step"><?php esc_html_e( 'Your details', 'woocommerce' ); ?></li>
+			<li class="woocommerce-order-withdrawal-content__step"><?php esc_html_e( 'Review and confirm', 'woocommerce' ); ?></li>
 		</ol>
 
-		<form class="woocommerce-OrderWithdrawalForm" method="post" novalidate>
+		<form class="woocommerce-OrderWithdrawalForm" action="<?php echo esc_url( $form_action_url ); ?>" method="post" novalidate>
 			<?php foreach ( $form_field_keys as $field_key ) : ?>
 				<?php
-				$field         = $fields[ $field_key ] ?? array();
-				$field_name    = isset( $field['name'] ) ? (string) $field['name'] : '';
-				$field_id      = isset( $field['id'] ) ? (string) $field['id'] : $field_name;
-				$field_type    = isset( $field['type'] ) ? (string) $field['type'] : 'text';
-				$field_classes = array_merge( array( 'woocommerce-form-row', 'form-row' ), (array) ( $field['classes'] ?? array() ) );
-				$description   = isset( $field['description'] ) ? (string) $field['description'] : '';
+				$field      = $fields[ $field_key ] ?? array();
+				$field_name = isset( $field['name'] ) ? (string) $field['name'] : '';
 
-				if ( isset( $form_errors[ $field_key ] ) ) {
-					$field_classes[] = 'woocommerce-invalid';
+				if ( '' === $field_name ) {
+					continue;
 				}
 				?>
-				<p class="<?php echo esc_attr( implode( ' ', $field_classes ) ); ?>">
-					<label for="<?php echo esc_attr( $field_id ); ?>"><?php echo esc_html( $field['label'] ?? '' ); ?>&nbsp;<span class="required" aria-hidden="true">*</span></label>
-					<input type="<?php echo esc_attr( $field_type ); ?>" class="woocommerce-Input woocommerce-Input--<?php echo esc_attr( $field_type ); ?> input-text" name="<?php echo esc_attr( $field_name ); ?>" id="<?php echo esc_attr( $field_id ); ?>" value="<?php echo esc_attr( $data[ $field_key ] ?? '' ); ?>" aria-required="true" <?php echo isset( $form_errors[ $field_key ] ) ? 'aria-invalid="true"' : 'aria-invalid="false"'; ?> <?php echo ! empty( $field['autocomplete'] ) ? 'autocomplete="' . esc_attr( (string) $field['autocomplete'] ) . '"' : ''; ?> <?php echo '' !== $description ? 'aria-describedby="' . esc_attr( $field_id . '_description' ) . '"' : ''; ?> />
-					<?php if ( '' !== $description ) : ?>
-						<span id="<?php echo esc_attr( $field_id . '_description' ); ?>" class="woocommerce-order-withdrawal-content__field-description"><?php echo esc_html( $description ); ?></span>
-					<?php endif; ?>
-				</p>
+				<?php woocommerce_form_field( $field_name, $field, $data[ $field_key ] ?? '' ); ?>
+
+				<?php if ( isset( $form_errors[ $field_key ] ) ) : ?>
+					<span id="<?php echo esc_attr( $field_name . '_error' ); ?>" class="woocommerce-order-withdrawal-content__field-error"><?php echo esc_html( $form_errors[ $field_key ] ); ?></span>
+				<?php endif; ?>
 
 				<?php if ( 'last_name' === $field_key ) : ?>
 					<div class="clear"></div>
@@ -146,26 +141,38 @@ $secondary_button_class = implode( ' ', array_merge( $button_classes, array( 'wo
 			if ( isset( $form_errors['withdrawal_type'] ) ) {
 				$withdrawal_type_classes[] = 'woocommerce-invalid';
 			}
+
+			$withdrawal_type_error_attributes = array(
+				'aria-invalid' => 'false',
+			);
+
+			if ( isset( $form_errors['withdrawal_type'] ) ) {
+				$withdrawal_type_error_attributes = array(
+					'aria-invalid'      => 'true',
+					'aria-errormessage' => $withdrawal_type_name . '_error',
+				);
+			}
 			?>
 			<fieldset class="<?php echo esc_attr( implode( ' ', $withdrawal_type_classes ) ); ?>">
 				<legend><?php echo esc_html( $withdrawal_type_field['label'] ?? '' ); ?>&nbsp;<span class="required" aria-hidden="true">*</span></legend>
 				<?php foreach ( $withdrawal_type_options as $option_value => $option_label ) : ?>
 					<label for="<?php echo esc_attr( $withdrawal_type_id . '_' . $option_value ); ?>" class="woocommerce-order-withdrawal-content__radio-label">
-						<input type="radio" class="input-radio" name="<?php echo esc_attr( $withdrawal_type_name ); ?>" id="<?php echo esc_attr( $withdrawal_type_id . '_' . $option_value ); ?>" value="<?php echo esc_attr( $option_value ); ?>" <?php checked( $data['withdrawal_type'] ?? '', $option_value ); ?> aria-required="true" <?php echo isset( $form_errors['withdrawal_type'] ) ? 'aria-invalid="true"' : 'aria-invalid="false"'; ?> />
+						<input type="radio" class="input-radio" name="<?php echo esc_attr( $withdrawal_type_name ); ?>" id="<?php echo esc_attr( $withdrawal_type_id . '_' . $option_value ); ?>" value="<?php echo esc_attr( $option_value ); ?>" <?php checked( $data['withdrawal_type'] ?? '', $option_value ); ?> aria-required="true" <?php echo wc_implode_html_attributes( $withdrawal_type_error_attributes ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?> />
 						<?php echo esc_html( $option_label ); ?>
 					</label>
 				<?php endforeach; ?>
 			</fieldset>
+			<?php if ( isset( $form_errors['withdrawal_type'] ) ) : ?>
+				<span id="<?php echo esc_attr( $withdrawal_type_name . '_error' ); ?>" class="woocommerce-order-withdrawal-content__field-error"><?php echo esc_html( $form_errors['withdrawal_type'] ); ?></span>
+			<?php endif; ?>
 
-			<?php
-			$additional_details_name = isset( $additional_details['name'] ) ? (string) $additional_details['name'] : '';
-			$additional_details_id   = isset( $additional_details['id'] ) ? (string) $additional_details['id'] : $additional_details_name;
-			?>
-			<p class="woocommerce-form-row form-row form-row-wide">
-				<label for="<?php echo esc_attr( $additional_details_id ); ?>"><?php echo esc_html( $additional_details['label'] ?? '' ); ?></label>
-				<textarea class="woocommerce-Input woocommerce-Input--textarea input-text" name="<?php echo esc_attr( $additional_details_name ); ?>" id="<?php echo esc_attr( $additional_details_id ); ?>" rows="5" aria-describedby="<?php echo esc_attr( $additional_details_id . '_description' ); ?>"><?php echo esc_textarea( $data['additional_details'] ?? '' ); ?></textarea>
-				<span id="<?php echo esc_attr( $additional_details_id . '_description' ); ?>" class="woocommerce-order-withdrawal-content__field-description"><?php echo esc_html( $additional_details['description'] ?? '' ); ?></span>
-			</p>
+			<?php if ( ! empty( $additional_details['name'] ) ) : ?>
+				<?php woocommerce_form_field( (string) $additional_details['name'], $additional_details, $data['additional_details'] ?? '' ); ?>
+
+				<?php if ( isset( $form_errors['additional_details'] ) ) : ?>
+					<span id="<?php echo esc_attr( (string) $additional_details['name'] . '_error' ); ?>" class="woocommerce-order-withdrawal-content__field-error"><?php echo esc_html( $form_errors['additional_details'] ); ?></span>
+				<?php endif; ?>
+			<?php endif; ?>
 
 			<p class="woocommerce-order-withdrawal-content__actions">
 				<?php wp_nonce_field( $nonce_action, $nonce_field ); ?>
